add status
status.cgi provides the current peak and RMS level for both, left and
right channel of the input and output streams.

status.html shows the status and updates it frequently. It can
optionally be moved to another server. For this purpose  the url in
status.cgi has to be adopted.